Move datatype inference to own file
