+++ test/golden/PrettyPrint/40b2e31ceea2ed197ad70a44de24646d50958007a8adc330683245a0d0176891.mimsa
@@ -1,0 +1,1 @@
+type Something    = String in True
