+++ test/golden/StoreExpr/40b2e31ceea2ed197ad70a44de24646d50958007a8adc330683245a0d0176891.json
@@ -1,0 +1,1 @@
+{"storeExpression":{"tag":"MyData","contents":[[],{"dtConstructors":{"String":[]},"dtName":"Something","dtVars":[]},{"tag":"MyLiteral","contents":[[],{"tag":"MyBool","contents":true}]}]},"storeBindings":{},"storeTypeBindings":{}}
